+++ packages/stream_chat_flutter/lib/src/components/avatar/stream_user_avatar.dart
@@ -155,6 +155,9 @@ class StreamUserAvatar extends StatelessWidget {
     .lg => StreamOnlineIndicatorSize.lg,
     .xl => StreamOnlineIndicatorSize.xl,
     .xxl => StreamOnlineIndicatorSize.xxl,
+    .xxxl => StreamOnlineIndicatorSize.xxxl,
+    // ignore: unreachable_switch_case, for forward compatibility
+    _ => StreamOnlineIndicatorSize.xxxl,
   };
 }
 
@@ -184,8 +187,10 @@ class _StreamUserAvatarPlaceholder extends StatelessWidget {
     final Widget content;
     if (userInitials != null && userInitials.isNotEmpty) {
       content = switch (size) {
-        .md || .lg || .xl || .xxl => Text(userInitials),
+        .md || .lg || .xl || .xxl || .xxxl => Text(userInitials),
         .xs || .sm => Text(userInitials.characters.first),
+        // ignore: unreachable_switch_case, for forward compatibility
+        _ => Text(userInitials),
       };
     } else {
       content = Icon(context.streamIcons.user);